TIP-221: move Product events
